Improved sync operation by adding --ignore-resources and --ignore-jobs options

+++ oar_utils/db/operations.py
@@ -2,8 +2,9 @@
 from __future__ import division, absolute_import, unicode_literals
 
 from copy import copy
+from functools import partial
 
-from sqlalchemy import func, MetaData, Table
+from sqlalchemy import func, MetaData, Table, and_, not_
 from sqlalchemy.engine.reflection import Inspector
 from sqlalchemy.types import Integer
 from sqlalchemy.sql.expression import select
@@ -16,11 +17,7 @@ from oar.lib.compat import  itervalues, to_unicode
 from .helpers import green, magenta, yellow, blue, red
 
 
-class NotSupportedDatabase(Exception):
-    pass
-
-
-UNUSED_TABLES = (
+UNUSED_TABLES = [
     'accounting',
     'gantt_jobs_predictions',
     'gantt_jobs_predictions_log',
@@ -28,7 +25,36 @@ UNUSED_TABLES = (
     'gantt_jobs_resources',
     'gantt_jobs_resources_log',
     'gantt_jobs_resources_visu',
-)
+]
+
+JOBS_TABLES = [
+    {'challenges': 'job_id'},
+    {'event_logs': 'job_id'},
+    {'frag_jobs': 'frag_id_job'},
+    {'job_dependencies': 'job_id'},
+    {'job_dependencies': 'job_id_required'},
+    {'job_state_logs': 'job_id'},
+    {'job_types': 'job_id'},
+    {'jobs': 'job_id'},
+    {'moldable_job_descriptions': 'moldable_job_id'},
+]
+
+MOLDABLE_JOBS_TABLES = [
+    {'assigned_resources': 'moldable_job_id'},
+    {'job_resource_groups': 'res_group_moldable_id'},
+    {'moldable_job_descriptions': 'moldable_id'},
+]
+
+jobs_table = [d.keys()[0] for d in JOBS_TABLES]
+moldable_jobs_tables = [d.keys()[0] for d in MOLDABLE_JOBS_TABLES]
+
+
+def get_table_columns(tables, table_name):
+    return [d[table_name] for d in tables if table_name in d.keys()]
+
+
+get_jobs_columns = partial(get_table_columns, JOBS_TABLES)
+get_moldables_columns = partial(get_table_columns, MOLDABLE_JOBS_TABLES)
 
 
 def copy_db(ctx):
@@ -36,11 +62,10 @@ def copy_db(ctx):
     if (not database_exists(engine_url) and is_local_database(ctx, engine_url)
         and ctx.current_db.dialect in ("postgresql", "mysql")):
             clone_db(ctx)
-    else:
-        tables = sync_schema(ctx)
-        sync_tables(ctx, tables)
-        if ctx.current_db.dialect == "postgresql":
-            fix_sequences(ctx)
+    tables = sync_schema(ctx)
+    sync_tables(ctx, tables, delete=True)
+    if ctx.current_db.dialect == "postgresql":
+        fix_sequences(ctx)
 
 
 def clone_db(ctx):
@@ -112,26 +137,43 @@ def sync_schema(ctx):
             yield Table(table.name, metadata, autoload=True)
 
 
-def sync_tables(ctx, tables):
+def get_sync_criteria(ctx, table):
+    # prepare query
+    criteria = []
+    if table.name in jobs_table:
+        for column_name in get_jobs_columns(table.name):
+            column = table.c.get(column_name)
+            criteria.append(column < ctx.max_job_to_sync)
+    if table.name in moldable_jobs_tables:
+        for column_name in get_moldables_columns(table.name):
+            column = table.c.get(column_name)
+            criteria.append(column < ctx.max_moldable_job_to_sync)
+    return criteria
+
+def sync_tables(ctx, tables, delete=False):
     # prepare the connection
     raw_conn = ctx.archive_db.engine.connect()
     # Get the max pk
     for table in tables:
         if table.name not in UNUSED_TABLES:
-            if table.primary_key:
-                pk = table.primary_key.columns.values()[0]
-                if isinstance(pk.type, Integer):
-                    max_pk_query = select([func.max(pk)])
-                    max_pk = raw_conn.execute(max_pk_query).scalar()
-                    cond = None
-                    if max_pk is not None:
-                        cond = (pk > max_pk)
-                    copy_table(ctx, table, raw_conn, condition=cond)
-                else:
-                    merge_table(ctx, table)
+            criteria = get_sync_criteria(ctx, table)
+            if delete and criteria:
+                reverse_criteria = [not_(c) for c in criteria]
+                delete_from_table(ctx, table, raw_conn, reverse_criteria)
             else:
-                delete_from_tables(ctx, table, raw_conn)
-                copy_table(ctx, table, raw_conn)
+                if table.primary_key:
+                    pk = table.primary_key.columns.values()[0]
+                    if isinstance(pk.type, Integer):
+                        max_pk_query = select([func.max(pk)])
+                        max_pk = raw_conn.execute(max_pk_query).scalar()
+                        if max_pk is not None:
+                            criteria.append(pk > max_pk)
+                        copy_table(ctx, table, raw_conn, criteria)
+                    else:
+                        merge_table(ctx, table)
+                else:
+                    delete_from_table(ctx, table, raw_conn)
+                    copy_table(ctx, table, raw_conn)
 
 
 def merge_table(ctx, table):
@@ -148,26 +190,31 @@ def merge_table(ctx, table):
     session.commit()
 
 
-def delete_from_tables(ctx, table, raw_conn, condition=None):
-    delete_query = table.delete()
-    if condition is not None:
-        ctx.log(magenta(' delete') + ' ~> from table ' + table.name)
-        delete_query = delete_query.where(condition)
-    else:
-        ctx.log(magenta('  empty') + ' ~> table ' + table.name)
-    raw_conn.execute(delete_query)
+def delete_from_table(ctx, table, raw_conn, criteria=[]):
+    check_query = select([func.count()]).select_from(table)
+    if criteria:
+        check_query = check_query.where(reduce(and_, criteria))
+    count = raw_conn.execute(check_query).scalar()
+    if count > 0:
+        delete_query = table.delete()
+        count_str = blue("%s/%s" % (count, count))
+        ctx.log(magenta(' delete') + ' ~> table %s (%s)' % (table.name, 
+                                                            count_str))
+        if criteria:
+            delete_query = delete_query.where(reduce(and_, criteria))
+        raw_conn.execute(delete_query)
 
 
-def copy_table(ctx, table, raw_conn, condition=None):
+def copy_table(ctx, table, raw_conn, criteria=[]):
     # prepare the connection
     from_conn = ctx.current_db.engine.connect()
 
     insert_query = table.insert()
     select_table = select([table])
     select_count = select([func.count()]).select_from(table)
-    if condition is not None:
-        select_query = select_table.where(condition)
-        count_query = select_count.where(condition)
+    if criteria:
+        select_query = select_table.where(reduce(and_, criteria))
+        count_query = select_count.where(reduce(and_, criteria))
     else:
         select_query = select_table
         count_query = select_count
@@ -175,9 +222,9 @@ def copy_table(ctx, table, raw_conn, condition=None):
     total_lenght = from_conn.execute(count_query).scalar()
     result = from_conn.execution_options(stream_results=True)\
                             .execute(select_query)
-    message = yellow('\r   copy') + ' ~> table %s (%s)'
-    ctx.log(message % (table.name, blue("0/%s" % total_lenght)), nl=False)
     if total_lenght > 0:
+        message = yellow('\r   copy') + ' ~> table %s (%s)'
+        ctx.log(message % (table.name, blue("0/%s" % total_lenght)), nl=False)
         progress = 0
         while True:
             transaction = raw_conn.begin()
@@ -191,7 +238,7 @@ def copy_table(ctx, table, raw_conn, condition=None):
             raw_conn.execute(insert_query, rows)
             del rows
         transaction.commit()
-    ctx.log("")
+        ctx.log("")
 
 
 def fix_sequences(ctx):
@@ -226,3 +273,11 @@ def is_local_database(ctx, engine_url):
     url = copy(engine_url)
     url.database = ctx.current_db.engine.url.database
     return url == ctx.current_db.engine.url
+
+
+class NotSupportedDatabase(Exception):
+    pass
+
+
+def purge_db():
+    pass
